# perf(thrift): use kitex BinaryProtocol replace apache BinaryProtocol for apache thrift codec