… go.sum
… pkg/remote/codec/thrift/deprecated.go