#### What type of PR is this?
<!--
Add one of the following kinds:

build: Changes that affect the build system or external dependencies (example scopes: gulp, broccoli, npm)
ci: Changes to our CI configuration files and scripts (example scopes: Travis, Circle, BrowserStack, SauceLabs)
docs: Documentation only changes
feat: A new feature
optimize: A new optimization
fix: A bug fix
perf: A code change that improves performance
refactor: A code change that neither fixes a bug nor adds a feature
style: Changes that do not affect the meaning of the code (white space, formatting, missing semi-colons, etc)
test: Adding missing tests or correcting existing tests
chore: Changes to the build process or auxiliary tools and libraries such as documentation generation
-->

#### Check the PR title.
<!--
The description of the title will be attached in Release Notes, 
so please describe it from user-oriented, what this PR does / why we need it.
Please check your PR title with the below requirements:
-->
- [ ] This PR title match the format: \<type\>(optional scope): \<description\>
- [ ] The description of this PR title is user-oriented and clear enough for others to understand.
- [ ] Attach the PR updating the user documentation if the current PR requires user awareness at the usage level. [User docs repo](https://github.com/cloudwego/cloudwego.github.io)


#### (Optional) Translate the PR title into Chinese.
perf(thrift): 使用 kitex BinaryProtocol 替换 apache BinaryProtocol 优化  apache thrift codec 性能

#### (Optional) More detailed description for this PR(en: English/zh: Chinese).
<!--
Provide more detailed info for review(e.g., it's recommended to provide perf data if this is a perf type PR).
-->
Note: old version always use kitex BinaryProtocol instead of apache, the perf loss was introduced in recent change(#1483) 
But this optimization still has a slight degradation compared to the previous version when using the Thrift buffered protocol
Thrift use apache thrift codec: (Kitex FastCodec has better performance)
v0.9.1: TPS: 212231.49, AVG: 0.47ms, TP99: 1.08ms, TP999: 1.96ms
#1490: TPS: 140546.15, AVG: 0.71ms, TP99: 1.42ms, TP999: 2.50ms (Use apache thrift BinaryProtocol directly)
this PR: TPS: 207607.18, AVG: 0.48ms, TP99: 1.01ms, TP999: 1.66ms (Replace apache thrift BinaryProtocol with Kitex BinaryProtocol)

#### (Optional) Which issue(s) this PR fixes:
<!--
Automatically closes linked issue when PR is merged.
Eg: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

#### (optional) The PR that updates user documentation:
<!--
If the current PR requires user awareness at the usage level, please submit a PR to update user docs. [User docs repo](https://github.com/cloudwego/cloudwego.github.io)
-->